forwarder-raw refactor (split into modules)
