+++ contracts/feature-tests/composability/forwarder-raw/src/forwarder_raw_sync.rs
@@ -1,0 +1,96 @@
+multiversx_sc::imports!();
+
+#[multiversx_sc::module]
+pub trait ForwarderRawSync: super::forwarder_raw_common::ForwarderRawCommon {
+    #[endpoint]
+    #[payable("EGLD")]
+    fn call_execute_on_dest_context(
+        &self,
+        to: ManagedAddress,
+        endpoint_name: ManagedBuffer,
+        args: MultiValueEncoded<ManagedBuffer>,
+    ) {
+        let payment = self.call_value().egld_value();
+        let half_gas = self.blockchain().get_gas_left() / 2;
+        let result = self.send_raw().execute_on_dest_context_raw(
+            half_gas,
+            &to,
+            &payment,
+            &endpoint_name,
+            &args.to_arg_buffer(),
+        );
+
+        self.execute_on_dest_context_result(result);
+    }
+
+    #[endpoint]
+    #[payable("EGLD")]
+    fn call_execute_on_dest_context_twice(
+        &self,
+        to: ManagedAddress,
+        endpoint_name: ManagedBuffer,
+        args: MultiValueEncoded<ManagedBuffer>,
+    ) {
+        let payment = self.call_value().egld_value();
+        let one_third_gas = self.blockchain().get_gas_left() / 3;
+        let half_payment = &*payment / 2u32;
+        let arg_buffer = args.to_arg_buffer();
+
+        let result = self.send_raw().execute_on_dest_context_raw(
+            one_third_gas,
+            &to,
+            &half_payment,
+            &endpoint_name,
+            &arg_buffer,
+        );
+        self.execute_on_dest_context_result(result);
+
+        let result = self.send_raw().execute_on_dest_context_raw(
+            one_third_gas,
+            &to,
+            &half_payment,
+            &endpoint_name,
+            &arg_buffer,
+        );
+        self.execute_on_dest_context_result(result);
+    }
+
+    #[endpoint]
+    #[payable("EGLD")]
+    fn call_execute_on_same_context(
+        &self,
+        to: ManagedAddress,
+        endpoint_name: ManagedBuffer,
+        args: MultiValueEncoded<ManagedBuffer>,
+    ) {
+        let payment = self.call_value().egld_value();
+        let half_gas = self.blockchain().get_gas_left() / 2;
+        let result = self.send_raw().execute_on_same_context_raw(
+            half_gas,
+            &to,
+            &payment,
+            &endpoint_name,
+            &args.to_arg_buffer(),
+        );
+
+        self.execute_on_same_context_result(result);
+    }
+
+    #[endpoint]
+    fn call_execute_on_dest_context_readonly(
+        &self,
+        to: ManagedAddress,
+        endpoint_name: ManagedBuffer,
+        args: MultiValueEncoded<ManagedBuffer>,
+    ) {
+        let half_gas = self.blockchain().get_gas_left() / 2;
+        let result = self.send_raw().execute_on_dest_context_readonly_raw(
+            half_gas,
+            &to,
+            &endpoint_name,
+            &args.to_arg_buffer(),
+        );
+
+        self.execute_on_dest_context_result(result);
+    }
+}
